Population.java
implemnt some new methods to store the following things:
- currentRound
- maxRounds
- averageFitness of the current population
- bestFitness the fitness of the current best individual

All individuals have now an object of these class
It is/was nessacary for the new fitness calculation that can be enabled

Individual.java
implement the method getDefaultFitness() it calculates the default fitness value:
		for (Car c : cars) {

			fitness = fitness + c.getUsedDuration();

		}
		fitness = fitness - period.getMaxServiceDurationPerDay()[0];
the modified and out commented getFitness() method show the new fitness. it is not the final fitness calculation and very slow so it is not implemnted yet.

Algorithm.java
the individual C (child) now get the population object from the first parent individual. Also implemented the Population class

Period.java
added the totalDemand